requirements.txt
alpaca-py
pandas
python-dotenv
matplotlib
